Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
astunparse==1.6.3
bandit==1.8.3
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==3.4.8
dbus-python==1.2.18
Deprecated==1.2.18
dill==0.4.0
distro==1.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
fire==0.7.0
flake8==7.2.0
Flask==3.1.1
github-webhook==1.0.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.10
importlib-metadata==4.6.4
isort==6.0.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.9.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==8.10.0
mypy_extensions==1.1.0
netifaces==0.11.0
oauthlib==3.2.0
openai==1.78.0
packaging==25.0
pathspec==0.12.1
pbr==6.1.1
platformdirs==4.3.8
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
PyGithub==2.6.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.10.1
pylint==3.3.7
PyNaCl==1.5.0
pyparsing==2.4.7
python-fire==0.1.0
python-gitlab==5.6.0
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
starlette==0.46.2
stevedore==5.4.1
termcolor==3.1.0
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wadllib==1.3.6
Werkzeug==3.1.3
wrapt==1.17.2
zipp==1.0.0
